Chlorophyll-a data from remote-sensing products

The CHL satellite data (mg/m$^3$) were downloaded from the NOAA ERDDAP server using Roy Mendelssohn's rerddapXtracto R package which uses the ropenscirerddap R packages available on CRAN. The R code used to download the data is in the extdata/get_satelite_data folder. See examples for how to find the file.

chl

Format

A data frame with:

Year

The year

Month

The month

CHL.1 to CHL.13

Average monthly CHL averaged over boxes 1 to 13. Units are mg/m$^3$.

Details

The Chlorophyll-a products are developed by the Ocean Biology Processing Group in the Ocean Ecology Laboratory at the NASA Goddard Space Flight Center.

For September 1997 to 2002, we used the Chlorophyll-a 2018.0 Reprocessing (R2018.0) product from the Sea-viewing Wide Field-of-view Sensor (SeaWiFS) on the Orbview-2 satellite. These data are on a 0.083 degree grid. See reference below.

For 2003 to 2017, we used the MODIS-Aqua R2018.1 product on a 0.0416 degree. These CHL data are taken from measurements gathered by the Moderate Resolution Imaging Spectroradiometer (MODIS) on NASA's Aqua Spacecraft. See reference below.

Both CHL data sets were downloaded from the NOAA ERDDAP server. See https://coastwatch.pfeg.noaa.gov/erddap/info/erdSW2018chlamday/index.html and https://coastwatch.pfeg.noaa.gov/erddap/info/erdMH1chlamday/index.html.

The R code used to download the data is in the extdata/get_satelite_data folder. See examples for how to find and view the files. The CHL values were averaged across thirteen 1 degree by 1 degree boxes which roughly parallel the bathymetry. See the figure in the examples below which shows the boxes.

References

NASA Goddard Space Flight Center, Ocean Ecology Laboratory, Ocean Biology Processing Group; (2018): SeaWiFS Ocean Color Data; NASA Goddard Space Flight Center, Ocean Ecology Laboratory, Ocean Biology Processing Group. https://dx.doi.org/10.5067/ORBVIEW-2/SEAWIFS/L3B/CHL/2018

NASA Goddard Space Flight Center, Ocean Ecology Laboratory, Ocean Biology Processing Group. Moderate-resolution Imaging Spectroradiometer (MODIS) Aqua Chlorophyll Data; 2014 Reprocessing. NASA OB.DAAC, Greenbelt, MD, USA. https://dx.doi.org/10.5067/AQUA/MODIS/L3M/CHL/2018

Hu C, Lee Z, Franz B (2012). “Chlorophyll aalgorithms for oligotrophic oceans: A novel approach based on three-band reflectance difference.” Journal of Geophysical Research: Oceans, 117(C1).

Chamberlain S (2019). rerddap: General Purpose Client for 'ERDDAP' Servers. R package version 0.6.5, https://CRAN.R-project.org/package=rerddap.

Mendelssohn R (2020). rerddapXtracto: Extracts Environmental Data from 'ERDDAP' Web Services. R package version 0.4.7, https://CRAN.R-project.org/package=rerddapXtracto.

Examples

if (FALSE) {
# Show the R code that downloaded the data
file.show(system.file("extdata/get_satellite_data", "get_sat_data.R", package="SardineForecast"))

# Show the background files on each data set at the time the data were downloaded
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdSW1chlamday", package="SardineForecast"))
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdMH1chlamday", package="SardineForecast"))

# Show the boxes
browseURL(system.file("docs", "kerala_study_area_with_inset.jpg", package="SardineForecast"))
}