Get R2 values for linear regression of log catch anomalies vs covariates

This function gets the R2 values for a linear regression of log catch anomaly ~ covariate anomalies at lags of 0 to 10. Covariates are turned into anomalies by removing year and qtr effect and then lagged and z-scored. Catch anomalies are logged and z-scored. Note: this is an old function and not used after I switched to using GAMs.

maker2qs(
  region = "Kerala",
  covname = "Precip",
  boxes = 1:14,
  rm.trend = TRUE,
  rm.season = TRUE
)

Arguments

region
region ("Kerala", "Karnataka", "Goa") in quotes.

covname
name of the covariate. Must be a colname in seio_covariates_qtr. If one of the covariates is from the boxes, then specify the part of the name in front of the number. For example "SST.".

boxes
for covariate from a box, the box number or numbers to use

rm.trend
remove year effect from catch and covariates

rm.season
remove season effect from catch and covariates

Value

A data.frame of R2's for each lag 0 to 10.

Examples

maker2qs(region="Kerala", covname="SST.", boxes=5)

#>               [,1]         [,2] [,3]
#>  [1,] 0.0032208429 -0.004002194  138
#>  [2,] 0.0005862301 -0.006708761  137
#>  [3,] 0.0009736737 -0.006372108  136
#>  [4,] 0.0043128856 -0.003062574  135
#>  [5,] 0.0192367537  0.011917625  134
#>  [6,] 0.0189546394  0.011578359  133
#>  [7,] 0.0362097860  0.028908345  132
#>  [8,] 0.0643423967  0.057199972  131
#>  [9,] 0.0707151346  0.063566790  130
#> [10,] 0.0577039341  0.050399313  129
#> [11,] 0.0344939089  0.026950893  128