Oil sardine landings in metric tons by quarter

A dataset containing the landings (in metric tons) of oil sardines in Kerala, Karnataka and Goa states 1956-2016. The data are collected and processed into a total landings estimates based on a stratified sampling of landing sites along the SW coast of India throughout the year. The program is run by the Central Marine Fisheries Research Institute (CMFRI) in Cochin, India. The data were obtained from reports published by CMFRI; see references.

oilsardine_qtr

Format

A data frame 'oilsardine_qtr' with 6 variables:

Year

Year

Qtr

Quarter 1 to 4

Goa

Goa landings

Karnataka

Karnataka landings

Kerala

Kerala landings

SWCoast

SW Coast landings

References

CMFRI reports were downloaded from the CMFRI Publication repository http://www.cmfri.org.in.

1956-1968 (Note Mysore which later becomes Karnataka) Antony Raja BT (1969). “Indian oil sardine.” CMFRI Bulletin, 16, 1--142.

1968-1978 Pillai VN (1982). Physical characteristics of the coastal waters off the south-west coast of India with an attempt to study the possible relationship with sardine, mackerel and anchovy fisheries. Ph.D. thesis, University of Cochin.

1975-1984 Kerala Jacob T, Rajendran V, Pillai PKM, Andrews J, Satyavan UK (1987). “An appraisal of the marine fisheries in Kerala.” Central Marine Fisheries Research Institute.

1975-1984 Kernataka Kurup KN, Nair GKK, Annam VP, Kant A, Beena MR, Kambadkar L (1987). “An appraisal of the marine fisheries of Karnataka and Goa.” Central Marine Fisheries Research Institute.

1985-2014 Provided by CMFRI directly via a data request.