Precipitation data

Three precipition datasets off the SW coast of India. Two are satellite derived and one is based on land gauges.

precip

Format

A data frame with:

Year

Month

precip.gpcp.kerala

Monthly GPCP precipitation in the box off the Kerala coast. Units are average mm/day.

precip.gpcp.karnataka

Monthly GPCP precipitation in the box off the Karnataka coast. Units are average mm/day.

PrecipTRMM

Monthly precipitation estimates from the Tropical Rainfall Measuring Mission (TRMM). Data are averaged in the boxes 1 to 13 used for the other satellite data.

Precip.Kerala

Monthly precipitation (mm) from landgauges.

Details

The National Climatic Data Center provides basic information on the GPCP Precipitation dataset. The dataset consists of monthly precipitation estimates (average mm/day) from January 1979 to the present. The precipitation estimates merge several satellite and in situ sources into a final product. Data are provided on a 2.5 degree grid. The GPCP Precipitation data are provided by the NOAA/NCEI Global Precipitation Climatology Project and were downloaded from https://www.ncei.noaa.gov/data/global-precipitation-climatology-project-gpcp-monthly. Two boxes were defined, one off the Kerala coast and one off the Karnataka coast, and the average values of all grid points within these boxes were used.

The boxes are 
Kerala Lat(8.75,  11.25), Lon(73.25,  75.75)
Karnataka Lat(13.75,  16.25), Lon(71.25,  73.75)

The land gauge data is monthly, seasonal and annual rainfall (in mm) area weighted average for each state in India starting from 1901 onwards. This data set is based on rain guages. The data are provided by the India Meteorological Department (Ministry of Earth Sciences). The 1901 to 2014 data were downloaded from the Open Government Data Platform India https://data.gov.in. The raw data are in the file 'Area_Weighted_Monthly_Seasonal_And_Annual_Rainfall_0.csv' in the extdata folder. The 2015 and 2016 data were extracted from the yearly Rainfall Statistics reports (see references).

NASA's Tropical Rainfall Measuring Mission (TRMM) website provides background on the TRMM data (https://pmm.nasa.gov/). 1997 to 2015 monthly precipitation estimates on a 0.25 degree grid were downloaded from the Tropical Rainfall Measuring Mission (TRMM) website. The data are averaged in the 2.5 x 2.5 degree boxes used for the other satellite data.

References

Adler RF, Huffman GJ, Chang A, Ferraro R, Xie P, Janowiak J, Rudolf B, Schneider U, Curtis S, Bolvin D, Gruber A, Susskind J, Arkin P (2003). “The Version 2 Global Precipitation Climatology Project (GPCP) Monthly Precipitation Analysis (1979-Present).” Journal of Hydrometeorology, 4, 1147--1167.

Adler R, Wang J, Sapiano M, Huffman G, Chiu L, Xie PP, Ferraro R, Schneider U, Becker A, Bolvin D, Nelkin E, Gu G, Program NOAACDR (2016). “Global Precipitation Climatology Project (GPCP) Climate Data Record (CDR), Version 2.3 (Monthly).” National Centers for Environmental Information. doi: 10.7289/V56971M6 , https://doi.org/doi:10.7289/V56971M6.

Purohit MK, Kaur S (2016). “Rainfall Statistics of India - 2016.” India Meterorological Department (Ministry of Earth Sciences). http://hydro.imd.gov.in/hydrometweb/.

Kothawale DR, Rajeevan M (2017). “Monthly, seasonal and annual rainfall time series for all-India, homogeneous regions and meteorological subdivisions: 1871-2016.” Indian Institute of Tropical Meteorology.

NCEI (2017). “Global Precipitation Climatology Project Monthly Product Version 2.3.” National Centers for Environmental Information. https://www.ncei.noaa.gov/data/global-precipitation-climatology-project-gpcp-monthly/access/.