Sea surface temperature data from remote-sensing products

The SST satellite data were downloaded from the NOAA ERDDAP server using Roy Mendelssohn's rerddapXtracto R package which uses the ropenscirerddap R package available on CRAN. The R code used to download the data is in the extdata/get_satelite_data folder. See examples for how to find the file. Units are degree Celsius.

sst

Format

A data frame with:

Year

The year

Month

The month

SST.1 to SST.13

Average monthly SST averaged over boxes 1 to 13. Units are degree Celcius.

SSTICOAD.1 to SSTICOAD.13

Average monthly SST averaged over boxes 1 to 13. Units are degree Celcius.

Details

There are two SST data sets: AVHRR and ICOADS.

AVHRR The SST differential upwelling indices were computed from the SST Daily Optimum Interpolation (OI), AVHRR Only, Version 2.1, Final 0.25 degree grid, Global data downloaded from the NOAA ERDDAP server. These are daily data. They daily data were averaged over all days in each month to produce monthly averages. The SST estimates use the Advanced Very-High Resolution Radiometer (AVHRR) instruments. AVHRR is accurate for close to the coast and thus could be used for our SST differential upwelling index.

The SST data were downloaded from the NOAA ERDDAP server. See https://coastwatch.pfeg.noaa.gov/erddap/info/ncdcOisst21Agg/index.html.

ICOADS The SST was extracted from the International Comprehensive Ocean-Atmosphere Data Set (ICOADS) collection of surface marine data Data from 1960 onward were used, which are on a 1°x1° grid.

The ICOADS SST data were downloaded from the NOAA ERDDAP server. https://coastwatch.pfeg.noaa.gov/erddap/info/esrlIcoads1ge/index.html More information on the ICOADS data are available at https://coastwatch.pfeg.noaa.gov/erddap/info/esrlIcoads1ge/index.html The R code used to download the data is in the extdata/get_satelite_data folder. See examples for how to find and view the files. The SST values were averaged across thirteen 1 degree by 1 degree boxes which roughly parallel the bathymetry.

References

The AVHRR data were provided by GHRSST and the US National Oceanographic Data Center. This project was supported in part by a grant from the NOAA Climate Data Record (CDR) Program for satellites. The data were downloaded from NOAA CoastWatch-West Coast Regional Node and Southwest Fisheries Science Center's Environmental Research Division. To cite these data in a paper, please follow the instructions in the license and at this link: https://coastwatch.pfeg.noaa.gov/erddap/information.html#citeDataset

The ICOADS data were provided by the NOAA/OAR/ESRL PSD, Boulder, Colorado, USA, from their Web site at http://www.esrl.noaa.gov/psd/ (and downloaded via NOAA's CoastWatch data server.)

Casey KS, Brandon TB, Cornillon P, Evans R (2010). “The past, present, and future of the AVHRR Pathfinder SST program.” In Oceanography from space, 273--287. Springer.

Walton CC, Pichel WG, Sapper JF, May DA (1998). “The development and operational application of nonlinear algorithms for the measurement of sea surface temperatures with the NOAA polar-orbiting environmental satellites.” Journal of Geophysical Research: Oceans, 103(C12), 27999--28012.

Chamberlain S (2019). rerddap: General Purpose Client for 'ERDDAP' Servers. R package version 0.6.5, https://CRAN.R-project.org/package=rerddap.

Mendelssohn R (2020). rerddapXtracto: Extracts Environmental Data from 'ERDDAP' Web Services. R package version 0.4.7, https://CRAN.R-project.org/package=rerddapXtracto.

Examples

if (FALSE) {
# Show the R code that downloaded the data
file.show(system.file("extdata/get_satellite_data", "get_sat_data.R", package="SardineForecast"))

# Show the background files on each data set at the time the data were downloaded
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdAGsstamday.html", package="SardineForecast"))
browseURL(system.file("extdata/get_satellite_data/ERDDAP_background", "erdPH2sstamday.html", package="SardineForecast"))

# Show the boxes
browseURL(system.file("docs", "kerala_study_area_with_inset.jpg", package="SardineForecast"))
}