Monthly tide height data from Cochin

Tide height data were downloaded form three sources to provide a 1939 to 2018 time series of tide height at Cochi. The Cochi tide gauge is on Willington Island; GLOSS# 032, UH#174, IOC#coch and PSMSL#438. The data are the raw tide gauge values, which are the mm above the datum for this gauge. The datum is 1.94m below BM27.

Sea level measurements are affected by the local barometric pressure, and it is common to apply the inverted barometer correction which uses the difference between the current air pressure and the mean air pressure. However Srinivas and Kumar (2006) found that for the Cochin tide height measurements, "There is no significant difference between the seasonal march of the observed sea level and CSL [corrected sea level]". Therefore we use the raw sea level data with no correction.

Srinivas et al. (2005) and Srinivas and Kumar (2006) found a positive correlation between the Southern Oscillation Index (SOI) and mean sea level for 1976-1993, with the sea level decreasing during El Nino years and increasing during La Nina years. Singh (2006) also found this correlation but for Sept only for Cochin. A Singh (2006) also found a correlation between the monthly rainfall in the SW monsoon season and mean sea level. Srinivas and Kumar (2006) also found a relationship between the along shore winds and sea level height at Cochi. Note that this relationship between sea level and SOI does not seem to hold for the entire 1939-2018 time series.

tide_mon

Format

tide_mon is a data frame with:

Year

The year

Month

The month

tide.level.mm

Tide level in mm above the Cochin datum (1.94m below BM27)

tide.level.mm

Tide level in mm above the Cochin datum (1.94m below BM27)

tide.level.interp

tide.level.mm with the NAs interpolated with exponential smoothing.

anomalies

tide.level.interp with linear trend and monthly means removed.

anomalies13

13 month running mean of anomalies as done in Srinivas et al. 2005 (Fig. 3).

anomaliesb

tide.level.interp with non-linear trend (via gam) and monthly means removed.

anomalies13b

13 month running mean of anomaliesb.

Details

For 1939 to 2013, mean monthly tide height was downloaded from the Permanent Service for Mean Sea Level (PSMSL), (PSMSL 2019, Holgate et al. 2013). This is from the tide gauge at Cochin on Willington Island (Lat 9.966667, Lon 76.266667). The data are the raw tide level data, termed the 'metric' data on the PSMSL website. Download link was https://www.psmsl.org/data/obtaining/met.monthly.data/438.metdata

In 2011, a automatic tide gauge was installed which provides minute tide height data. Data from 2011 onward are provided by Indian National Centre for Ocean Information Services (INCOIS) and hosted at http://www.ioc-sealevelmonitoring.org and the University of Hawaii Sea Level Center (UHSLC) https://uhslc.soest.hawaii.edu/. Research quality data are not available for Cochin (UH# 174). Instead the Fast Delivery data were used and quality control (outlier removal) was applied to the downloaded data per the algorithm used by IOC Sea Level Monitoring (FAQ http://www.ioc-sealevelmonitoring.org/service.php).

Daily minute tide level data were downloaded from http://www.ioc-sealevelmonitoring.org/station.php?code=coch and outlier removal applied.

All values X where abs(X – median) > tolerance are hidden.

With tolerance = 3*abs(percentile90 - median)

The statistics (median and percentile90) are calculated on the daily minute data

After outlier removal, NAs were replaced with a linear interpolation and the mean tide height for the day (average over all the minute data) was computed. The daily data were then averaged over the calendar month to create the monthly average. Months with fewer than 20 days of data were given NA. was computed.

Daily mean tide level was also downloaded from the University of Hawaii Sea Level Center (UHSLC) at the link http://uhslc.soest.hawaii.edu/data/?fd#uh174. This data should be the same as the IOC data above but there were slight differences that are likely due to differences in outlier removal and how NAs are treated. As for the IOC data, he daily data were then averaged over the calendar month to create the monthly average. Months with fewer than 20 days of data were given NA. was computed. The reference for these data are Caldwell et al. (2015).

A monthly time series for 1939 to 2018 was assembled by merging the 3 data sets. The PSMSL data were used for 1939 to Aug 2011. For Sept 2011 onward, the IOC and UHSLC monthly data were averaged. If one had a NA and the other did not, the non-NA value was used. The monthly data had NAs. The NAs were interpolated using exponential smoothing (ETS) via a seasonal model.

An ETS model was fit to the data up to the first NA

The fitted ETS model was used to forecast one step ahead.

The one-step ahead forecast was used to fill in the first NA

A new ETS model was fit to the data up to the second NA, where the first NA was filled in via the step above.

The new fitted ETS model was used to forecast one step ahead.

This one-step ahead forecast was used to fill in the second NA.

This process was repeated until all NAs were replaced.

The R code for interpolation in the 'process-merged-tide-gauge-data.R' file.

All code for downloading and processing the data along with the raw downloaded data is in the extdata/tide gauge data folder in the doc folder.

References

Caldwell, P. C., M. A. Merrifield, P. R. Thompson (2015), Sea level measured by tide gauges from global oceans — the Joint Archive for Sea Level holdings (NCEI Accession 0019568), Version 5.5, NOAA National Centers for Environmental Information, Dataset, doi:10.7289/V5V40S7W. Research Quality data is not available for the Cochin tide guage (UH# 174, GLOSS# 032). The Fast Delivery data was downloaded on May 7, 2019. from http://uhslc.soest.hawaii.edu/data/?fd#uh174

Holgate SJ, Matthews A, Woodworth PL, Rickards LJ, Tamisiea ME, Bradshaw E, Foden PR, Gordon KM, Jevrejeva S, Pugh J (2013). “New data systems and products at the Permanent Service for Mean Sea Level.” Journal of Coastal Research, 29(3), 393--504. doi: 10.2112/JCOASTRES-D-12-00175.1. , https://doi.org/doi:10.2112/JCOASTRES-D-12-00175.1..

Permanent Service for Mean Sea Level (PSMSL). 2019. "Tide Gauge Data", Retrieved 6 May 2019 from https://www.psmsl.org/data/obtaining/stations/438.php.

Singh OP (2006). “ENSO and monsoon induced sea level changes and their impacts along the Indian coastline.” Indian Journal of Marine Sciences, 35(2), 87--92.

Srinivas K, Kumar PKD (2006). “Atmospheric forcing on the seasonal variability of sea level at Cochin, southwest coast of India.” Continental Shelf Research, 26(10), 1113--1133. doi: 10.1016/j.csr.2006.03.010. , https://doi.org/https://doi.org/10.1016/j.csr.2006.03.010..

Srinivas K, Kumar PKD, Revichandran C (2005). “ENSO signature in the sea level along the coastline of the Indian subcontinent.” Indian Journal of Marine Sciences, 34(2), 225--236.