Upwelling indices from remote-sensing products

Three upwelling indices are in the 'upw' data object: a SST nearshore offshore differential, a wind-based index, Ekman Mass Transport, and Ekman Pumping. The wind upwelling indices and SST data were downloaded from the NOAA ERDDAP server using R Mendels rerddapXtracto R package which uses the ropensci rerddap R package. The R code used to download the data is in the extdata/get_satelite_data folder. See examples for how to find the file. The Ekman Mass Transport and Pumping were computed from winds from the ERA5 product.

upw

Format

A data frame with:

Year

The year

Month

The month

Wind.UPW.1 to Wind.UPW.5

Average monthly wind-based upwelling index averaged over boxes 1 to 5. Units are m/s.

SST.UPW.1 to SST.UPW.5

Average monthly SST differential averaged over boxes 1 to 5. Units are degree Celcius.

Bakun.UPW

Bakun index at 74.5E 11.5N (near Kochi, India) at a coast angle of 158 degrees. Units are kg m$^-1$ s$^-1$.

Details

The wind-based monthly upwelling indices were downloaded from the NOAA ERDDAP server. The first is 1999-2009 on a 0.125 degree grid. The second is 2009 to present on a 0.25 degree grid. Units are m/s. See https://coastwatch.pfeg.noaa.gov/erddap/info/erdQSstressmday/index.html and https://coastwatch.pfeg.noaa.gov/erddap/info/erdQAstressmday/index.html.

The SST differential upwelling indices were computed from the SST Daily Optimum Interpolation (OI), AVHRR Only, Version 2.1, Final 0.25 degree grid, Global data downloaded from the NOAA ERDDAP server. This is AVHRR so accurate for close to the coast. These are daily data. They daily data was averaged over all days in each month to produce monthly averages. https://coastwatch.pfeg.noaa.gov/erddap/info/ncdcOisst21Agg/index.html. The UPW index is the difference between the coast box (1 to 5) and a box 3 degrees offshore at the same latitude.

The Bakun index (The Bakun 1973) is calculated based upon Ekman's theory of mass transport due to wind stress. The index is computed from the ektrx and ektry, which are the x- and y- components of Ekman Transport (kg m$^-1$ s$^-1$) obtained from the ERDDAP link below, and coast_angle is 158 degrees for the India west coast near Kochi (74.5E 11.5N coast angle 158 degrees). The Bakun index was downloaded from https://coastwatch.pfeg.noaa.gov/erddap/info/erdlasFnWPr/index.html. The function to compute the Bakun index is (from https://oceanview.pfeg.noaa.gov/products/upwelling/bakun)

 upwell <- function(ektrx, ektry, coast_angle) {
 pi <- 3.1415927
 degtorad <- pi/180.
 alpha <- (360 - coast_angle) * degtorad
 s1 <- cos(alpha)
 t1 <- sin(alpha)
 s2 <- -1 * t1
 t2 <- s1
 perp <- (s1 * ektrx) + (t1 * ektry)
 para <- (s2 * ektrx) + (t2 * ektry)
 return(perp/10)
 }
 

The R code used to download the data is in the extdata/get_satelite_data folder.

The wind data are from the ERA5 Reanalysis Dataset https://confluence.ecmwf.int/display/CKB/ERA5. downloaded from the Asia-Pacific Data-Research Center ERDDAP server https://apdrc.soest.hawaii.edu/erddap/info/hawaii_soest_66d3_10d8_0f3c/index.html.

The R code used to download the data is in extdata/get_satelite_data/get_EMTperp.R and the functions are in extdata/get_satelite_data/Replicate_EMT/get_EMT_functions.R.

References

SST data: These data were provided by GHRSST and the US National Oceanographic Data Center. This project was supported in part by a grant from the NOAA Climate Data Record (CDR) Program for satellites. The data were downloaded from NOAA CoastWatch-West Coast Regional Node and Southwest Fisheries Science Center's Environmental Research Division.

Wind-based UPW index: NOAA's CoastWatch Program distributes wind velocity measurements derived from the Seawinds instrument aboard NASA's QuikSCAT satellite. The Seawinds instrument is a dual-beam microwave scatterometer designed to measure wind magnitude and direction over the global oceans. CoastWatch further processes these wind velocity measurements to wind stress and wind stress curl.

Bakun index: The Environmental Research Division (ERD), within NOAA Fisheries, has long been a leader in development and calculation of upwelling and other environmental indices. ERD was originally established as the Pacific Environmental Group at the U.S. Navy Fleet Numerical Meteorology and Oceanography Center (FNMOC) in Monterey, California, to take advantage of the Navy's global oceanographic and meteorological models. FNMOC produces operational forecasts of the state of the atmosphere and the ocean several times daily. Before the advent of satellite oceanography, these forecasts provided global snapshots of ocean conditions for Navy operations, but were also invaluable for studies of fisheries climatology since they provided long time series of environmental conditions at a much higher resolution than was possible from direct measurement. The FNMOC sea-level pressure became the basis of the Bakun upwelling index calculation, and provides estimates of upwelling for the Northern Hemisphere starting in 1948 and globally since 1981.

Bakun A (1973). “Coastal upwelling indices, west coast of North America.” US Department of Commerce, NOAA Technical Report NMFS SSRF-671.

Examples

if (FALSE) {
# Show the R code that downloaded the data
file.show(system.file("extdata/get_satellite_data", "get_sat_data.R", package="SardineForecast"))

# Show the R code that downloaded the data
file.show(system.file("extdata/get_satellite_data", "get_bakun_upi_data.R", package="SardineForecast"))

# Show the boxes
browseURL(system.file("docs", "kerala_study_area_with_inset.jpg", package="SardineForecast"))
}